Separate the FA, RCA to individual module
