+++ lib/rohd_hcl.dart
@@ -5,7 +5,9 @@ export 'src/arbiter.dart';
 export 'src/carry_save_multiplier.dart';
 export 'src/exceptions.dart';
 export 'src/fifo.dart';
+export 'src/full_adder.dart';
 export 'src/interfaces/interfaces.dart';
 export 'src/memory.dart';
+export 'src/ripple_carry_adder.dart';
 export 'src/one_hot.dart';
 export 'src/rotate.dart';
