+++ lib/src/full_adder.dart
@@ -1,0 +1,41 @@
+// Copyright (C) 2023 Intel Corporation
+// SPDX-License-Identifier: BSD-3-Clause
+//
+// carry_save_multiplier.dart
+// Implementation of pipeline multiplier module.
+//
+// 2023 May 29
+// Author: Yao Jing Quek <yao.jing.quek@intel.com>
+//
+
+import 'package:rohd/rohd.dart';
+
+/// A simple full-adder with inputs `a` and `b` to be added with a `carryIn`.
+class FullAdder extends Module {
+  /// The result [sum] from [FullAdder].
+  Logic get sum => output('sum');
+
+  /// The result [cOut] from [FullAdder].
+  Logic get cOut => output('carry_out');
+
+  /// Constructs a [FullAdder] with value [a], [b] and [carryIn].
+  FullAdder({
+    required Logic a,
+    required Logic b,
+    required Logic carryIn,
+    super.name = 'full_adder',
+  }) {
+    a = addInput('a', a, width: a.width);
+    b = addInput('b', b, width: b.width);
+    carryIn = addInput('carry_in', carryIn, width: carryIn.width);
+
+    final carryOut = addOutput('carry_out');
+    final sum = addOutput('sum');
+
+    final and1 = carryIn & (a ^ b);
+    final and2 = b & a;
+
+    sum <= (a ^ b) ^ carryIn;
+    carryOut <= and1 | and2;
+  }
+}
